feat: moved dataBadge to other file
