+++ components/DataBadge/src/index.tsx
@@ -1,24 +1,6 @@
-import React from 'react';
-import clsx from 'clsx';
 import './index.scss';
-import {
-  DataBadge as NLDataBadge,
-  DataBadgeProps as NLDataBadgeProps,
-} from '@nl-design-system-candidate/data-badge-react';
-export * from '@nl-design-system-candidate/data-badge-react';
-import '@nl-design-system-candidate/data-badge-css/data-badge.css';
+import DataBadge from './DataBadge';
 
-export type DataBadgeProps = NLDataBadgeProps & {
-  status?: 'neutral' | 'success' | 'warning' | 'error';
-};
+export default DataBadge;
 
-export const DataBadge = ({ status, ...props }: DataBadgeProps) => {
-  const classes = clsx({
-    'nl-data-badge--neutral': status === 'neutral',
-    'nl-data-badge--success': status === 'success',
-    'nl-data-badge--warning': status === 'warning',
-    'nl-data-badge--error': status === 'error',
-  });
-
-  return <NLDataBadge className={classes} {...props} />;
-};
+export * from './DataBadge';
