+++ components/Heading/src/index.tsx
@@ -1,2 +1,2 @@
-export * from '@nl-design-system-candidate/heading-react';
 import '@nl-design-system-candidate/heading-css/heading.css';
+export * from '@nl-design-system-candidate/heading-react';
