+++ components/NumberBadge/src/index.tsx
@@ -1,2 +1,2 @@
+import '@nl-design-system-candidate/number-badge-css/number-badge.css';
 export * from '@nl-design-system-candidate/number-badge-react';
-export * from '@nl-design-system-candidate/number-badge-react/css';
