+++ components/Paragraph/src/index.tsx
@@ -1,2 +1,2 @@
-export * from '@nl-design-system-candidate/paragraph-react';
 import '@nl-design-system-candidate/paragraph-css/paragraph.css';
+export * from '@nl-design-system-candidate/paragraph-react';
