+++ package.json
@@ -89,5 +89,5 @@
     "node": "^20",
     "pnpm": "^10"
   },
-  "packageManager": "pnpm@10.2.0+sha256.72f1630812107c8723eb5be4517725f3e13f09cfd8f71571cf29f23ce318bdb7"
+  "packageManager": "pnpm@10.9.0"
 }
